+++ api-module-library/ironclad/test/api.test.js
@@ -40,13 +40,14 @@ describe('Ironclad API class', () => {
 
         it('should delete a webhook', async () => {
             const response = await api.deleteWebhook(webhookID);
-            expect(response.status).equal(204);
+            expect(response.status).to.equal(204);
         });
     });
 
     describe('Workflows', () => {
         let workflowSchemaID;
         let workflowID;
+        let documentKey;
         it('should list all workflows', async () => {
             const response = await api.listAllWorkflows();
             expect(response).to.have.property('page');
@@ -54,6 +55,18 @@ describe('Ironclad API class', () => {
             expect(response).to.have.property('count');
             expect(response).to.have.property('list');
             workflowID = response.list[0].id;
+        });
+
+        it('should return the second page of workflows', async () => {
+            let params = {
+                page: 1,
+            };
+
+            const response = await api.listAllWorkflows(params);
+            expect(response).to.have.property('page');
+            expect(response).to.have.property('pageSize');
+            expect(response).to.have.property('count');
+            expect(response).to.have.property('list');
         });
 
         it('should create a workflow', async () => {
@@ -126,14 +139,90 @@ describe('Ironclad API class', () => {
             expect(response).to.have.property('approvals');
             expect(response).to.have.property('signatures');
             expect(response).to.have.property('isRevertibleToReview');
+            documentKey = response.attributes.draft[0].download.split('/')[7];
         });
 
-        it('should list all workflow approvals', async () => {
+        it.skip('should list all workflow approvals', async () => {
             const response = await api.listAllWorkflowApprovals(workflowID);
             expect(response).to.have.property('workflowId');
             expect(response).to.have.property('title');
             expect(response).to.have.property('approvalGroups');
             expect(response).to.have.property('roles');
+        });
+
+        it('should update a workflow approval', async () => {
+            workflowID = process.env.WORKFLOW_ID;
+            const workflowApprovals = await api.listAllWorkflowApprovals(workflowID);
+            workflowApprovals.approvalGroups.forEach(async (approvalGroup) => {
+                let roleID = approvalGroup.reviewers[0].role;
+                let role = workflowApprovals.roles.find((role) => role.id === roleID)
+                let email = role.assignees[0].email
+
+                let body = {
+                    user: {
+                        type: 'email',
+                        email
+                    },
+                    status: 'approved'
+                }
+
+                const response = await api.updateWorkflowApprovals(workflowID, roleID, body);
+                expect(response).to.equal(true)
+            })
+
+        it('should create a workflow comment', async () => {
+            const body = {
+                creator: {
+                    type: 'email',
+                    email: 'projectteam@lefthook.com',
+                },
+                comment: 'Testing a comment',
+            };
+
+            const response = await api.createWorkflowComment(workflowID, body);
+            expect(response).to.equal('');
+        });
+
+        it('should retrieve a workflow document', async () => {
+            const response = await api.retrieveWorkflowDocument(
+                workflowID,
+                documentKey
+            );
+            expect(response).to.exist;
+        });
+
+        // Must be workflow in review step
+        it('should update a workflow metadata', async () => {
+            const params = {
+                status: 'active',
+            };
+            const datetime = Date.now();
+            const getReviewWorkflows = await api.listAllWorkflows(params);
+            let reviewWorkflowId = null;
+
+            for (const workflow of getReviewWorkflows.list) {
+                if (workflow.step === 'Review') {
+                    reviewWorkflowId = workflow.id;
+                    break;
+                }
+            }
+
+            const body = {
+                // Testing string, also should test:
+                //    Email, Number, Boolean, Date, Dynamic table, Monetary value, Address, General object
+                updates: [
+                    {
+                        action: 'set',
+                        path: 'counterpartyName',
+                        value: `Updated Example Company ${datetime}`,
+                    },
+                ],
+                comment: 'Updated workflow counterpartyName',
+            };
+            const response = await api.updateWorkflow(reviewWorkflowId, body);
+            expect(response).to.have.property('id');
+            expect(response).to.have.property('title');
+            expect(response).to.have.property('schema');
         });
     });
 
@@ -160,6 +249,7 @@ describe('Ironclad API class', () => {
             expect(response).to.have.property('type');
             expect(response).to.have.property('name');
             expect(response).to.have.property('lastUpdated');
+            recordID = response.id;
         });
 
         it('should list all records', async () => {
@@ -168,7 +258,6 @@ describe('Ironclad API class', () => {
             expect(response).to.have.property('pageSize');
             expect(response).to.have.property('count');
             expect(response).to.have.property('list');
-            recordID = response.list[0].id;
         });
 
         it('should list all record schemas', async () => {